major changes

+++ components/MyClasses.js
@@ -1,32 +1,51 @@
-import { View, Text, StyleSheet, TouchableOpacity, FlatList } from 'react-native';
+import { View, Text, StyleSheet, TouchableOpacity, FlatList, Alert } from 'react-native';
 import React, { useState, useEffect } from 'react';
 import { MaterialIcons } from '@expo/vector-icons';
-import { useNavigation } from '@react-navigation/native';
+import { useNavigation,  useFocusEffect} from '@react-navigation/native';
 import Toolbar from './Toolbar';
 import { FontAwesome } from '@expo/vector-icons';
 import { auth, db } from './firebase';
 import { collection, query, where, getDocs, deleteDoc } from 'firebase/firestore';
 
+
 const MyClasses = (props) => {
   const navigation = useNavigation();
 
   const [data, setData] = useState([]);
-  const [students, setStudents] = useState([]);
+  const [ids, setIdes] = useState([]);
+  const [classesChanged, setClassesChanged] = useState(false);
 
+  useFocusEffect(
+    React.useCallback(() => {
+      setClassesChanged(true);
+    }, [])
+  );
+  
+  useEffect(() => {
+    if (classesChanged) {
+      getClasses().then(() => setClassesChanged(false)).catch(console.error);
+    }
+  }, [classesChanged]);
 
   const getClasses = async () => {
     const q = query(collection(db, 'Classes'), where('t_id', '==', auth.currentUser.uid));
     const querySnapshot = await getDocs(q);
     const classNames = [];
+    const classIds = [];
     querySnapshot.forEach((doc) => {
-      classNames.push(doc.id);
+      const classData = doc.data();
+      if (classData.class_name) {
+        classNames.push(classData.class_name);
+        console.log("hey")
+      }
+      classIds.push(doc.id); // Use doc.id instead of classData.class_id
     });
+    classNames.sort(); // Sort the classNames array alphabetically
     setData(classNames);
+    setIdes(classIds);
   };
-
-  useEffect(() => {
-    getClasses().catch(console.error);
-  }, []);
+  
+  
 
   const handleDeleteStudents = async (className) => {
     const q = query(collection(db, 'Students'), where('t_id', '==', auth.currentUser.uid), where('class_id', '==', className));
@@ -40,26 +59,48 @@ const MyClasses = (props) => {
   const handleDeleteClasses = async (className) => {
     const q = query(collection(db, 'Classes'), where('t_id', '==', auth.currentUser.uid), where('class_name', '==', className));
     const querySnapshot = await getDocs(q);
-    querySnapshot.forEach(async (doc) => {
-      await deleteDoc(doc.ref);
-    });
-    setData(prevData => prevData.filter(item => item !== className));
+    if (querySnapshot.empty) {
+      console.log('No matching documents.');
+      return;
+    }
+  
+    Alert.alert(
+      'Delete Class',
+      'Are you sure you want to delete this class?',
+      [
+        { text: 'No', onPress: () => console.log('Cancel Pressed'), style: 'cancel' },
+        { text: 'Yes', onPress: async () => {
+            querySnapshot.forEach(async (doc) => {
+              await deleteDoc(doc.ref);
+            });
+            setData(prevData => prevData.filter(item => item !== className));
+          }
+        },
+      ],
+      { cancelable: false }
+    );
   };
-  const renderItem = ({ item }) => (
+
+  const handleAddClass = () => {
+    setClassesChanged(true); 
+    navigation.navigate('SetDetails');
+  };
+  
+  const renderItem = ({ item, index }) => (
     <View style={{ flexDirection: 'row' }}>
-      <TouchableOpacity onPress={() => navigation.navigate('ClassDetails', { className: item })}>
+      <TouchableOpacity onPress={() => navigation.navigate('ClassDetails', { class_id: ids[index] , class_name: data[index]})}>
         <Text style={{ padding: 10, fontSize: 22, textAlign: 'center', textDecorationLine: 'underline', flexDirection: 'column', justifyContent: 'center' }}>
           {item}
         </Text>
+
       </TouchableOpacity>
       <TouchableOpacity onPress={() => handleDeleteStudents(item)}>
         <MaterialIcons name="delete" size={24} color="black" />
       </TouchableOpacity>
-
-      
-
     </View>
   );
+
+  
 
   return (
     <View style={styles.allPage}>
@@ -70,9 +111,9 @@ const MyClasses = (props) => {
       </View>
       
       <View>
-      <TouchableOpacity style={styles.button} onPress={() => navigation.navigate('SetDetails')}>
-    <Text style={styles.buttonText}>הוסף כיתה</Text>
-  </TouchableOpacity>
+      <TouchableOpacity style={styles.button} onPress={handleAddClass}>
+        <Text style={styles.buttonText}>הוסף כיתה</Text>
+      </TouchableOpacity>
         <FlatList
           data={data}
           renderItem={renderItem}
@@ -82,7 +123,7 @@ const MyClasses = (props) => {
 
       </View>
       
-      <TouchableOpacity style={[styles.back, { marginTop: 'auto' }]} onPress={() => navigation.navigate('HomePage')}>
+      <TouchableOpacity style={[styles.back, { marginTop: 'auto' }]} onPress={() => navigation.navigate('HomePage') }>
         <MaterialIcons name="navigate-next" size={24} color="black" />
         <Text>הקודם</Text>
       </TouchableOpacity>
